feat: Create TRPC Local Link (#2928)
* feat: create local link

* fix: tests
