+++ packages/web/server/api/root.ts
@@ -1,3 +1,4 @@
+import { localRouter } from "~/server/api/local-routers/local-router";
 import { bridgeTransferRouter } from "~/server/api/routers/bridge-transfer";
 import { createTRPCRouter } from "~/server/api/trpc";
 
@@ -12,6 +13,7 @@ import { edgeRouter } from "./edge-routers/edge-router";
 export const appRouter = createTRPCRouter({
   bridgeTransfer: bridgeTransferRouter,
   edge: edgeRouter,
+  local: localRouter,
 });
 
 export type AppRouter = typeof appRouter;
