+++ packages/web/utils/trpc.ts
@@ -1,12 +1,25 @@
 import "~/utils/superjson";
 
 import { logEvent } from "@amplitude/analytics-browser";
-import { httpBatchLink, httpLink, loggerLink, splitLink } from "@trpc/client";
+import {
+  httpBatchLink,
+  httpLink,
+  loggerLink,
+  OperationResultEnvelope,
+  splitLink,
+  TRPCClientError,
+  TRPCLink,
+} from "@trpc/client";
 import { createTRPCNext } from "@trpc/next";
-import { type inferRouterInputs, type inferRouterOutputs } from "@trpc/server";
+import {
+  AnyRouter,
+  type inferRouterInputs,
+  type inferRouterOutputs,
+} from "@trpc/server";
+import { observable } from "@trpc/server/observable";
 
 import { EventName } from "~/config";
-import { type AppRouter } from "~/server/api/root";
+import { type AppRouter, appRouter } from "~/server/api/root";
 import { superjson } from "~/utils/superjson";
 
 const getBaseUrl = () => {
@@ -14,6 +27,42 @@ const getBaseUrl = () => {
   if (process.env.VERCEL_URL) return `https://${process.env.VERCEL_URL}`; // SSR should use vercel url
   return `http://localhost:${process.env.PORT ?? 3000}`; // dev SSR should use localhost
 };
+
+/**
+ * Creates a local link for tRPC operations.
+ * This function is used to create a custom TRPCLink that intercepts operations and
+ * handles them locally using the provided router, instead of sending them over the network.
+ */
+export function localLink<TRouter extends AnyRouter>({
+  router,
+}: {
+  router: TRouter;
+}): TRPCLink<TRouter> {
+  return () =>
+    ({ op }) =>
+      observable<OperationResultEnvelope<unknown>, TRPCClientError<TRouter>>(
+        (observer) => {
+          async function execute() {
+            const caller = router.createCaller({});
+            try {
+              // Attempt to execute the operation using the router's caller.
+              const data = await (
+                caller[op.path] as (input: unknown) => unknown
+              )(op.input);
+              // If successful, notify the observer with the result.
+              observer.next({ result: { data, type: "data" } });
+              observer.complete();
+            } catch (err) {
+              // If an error occurs, convert it to a TRPCClientError and notify the observer.
+              observer.error(TRPCClientError.from(err as Error));
+            }
+          }
+
+          // Execute the operation asynchronously.
+          void execute();
+        }
+      );
+}
 
 /** Provides ability to skip batching given a new custom query option context: `skipBatch: boolean` */
 const makeSkipBatchLink = (url: string) =>
@@ -83,6 +132,7 @@ export const api = createTRPCNext<AppRouter>({
           const servers = {
             node: makeSkipBatchLink(`${getBaseUrl()}/api/trpc`)(runtime),
             edge: makeSkipBatchLink(`${getBaseUrl()}/api/edge-trpc`)(runtime),
+            local: localLink({ router: appRouter })(runtime),
 
             /**
              * Create a separate link for the pools edge server since its query is too expensive
@@ -109,11 +159,14 @@ export const api = createTRPCNext<AppRouter>({
              * If the base path is not `edge`, we can just call the node server directly.
              */
             const isEdge = basePath === "edge";
+            const isLocal = basePath === "local";
             const isPoolsEdge = isEdge && possibleEdgePath.startsWith("pools");
 
             let link: (typeof servers)["node"];
             if (isEdge && !isPoolsEdge) {
               link = servers["edge"];
+            } else if (isLocal) {
+              link = servers["local"];
             } else if (isPoolsEdge && possibleEdgePath.startsWith("pools")) {
               link = servers["poolsEdge"];
             } else {
